refactoring for shorter test function names
